+++ src/modules/sbstudio/plugin/utils/sampling.py
@@ -8,7 +8,7 @@ from sbstudio.model.light_program import LightProgram
 from sbstudio.model.point import Point4D
 from sbstudio.model.trajectory import Trajectory
 from sbstudio.model.yaw import YawSetpoint, YawSetpointList
-from sbstudio.plugin.colors import get_color_of_drone
+from sbstudio.plugin.tasks.light_effects import get_final_color_of_drone
 from sbstudio.plugin.utils.evaluator import (
     get_position_of_object,
     get_xyz_euler_rotation_of_object,
@@ -205,7 +205,7 @@ def sample_colors_of_objects(
     for _, time in each_frame_in(frames, context=context, redraw=redraw):
         for obj in objects:
             key = obj.name
-            color = get_color_of_drone(obj)
+            color = get_final_color_of_drone(obj)
             lights[key].append(
                 Color4D(
                     time,
@@ -258,7 +258,7 @@ def sample_positions_and_colors_of_objects(
         for obj in objects:
             key = obj.name
             pos = get_position_of_object(obj)
-            color = get_color_of_drone(obj)
+            color = get_final_color_of_drone(obj)
             trajectories[key].append(Point4D(time, *pos))
             lights[key].append(
                 Color4D(
@@ -319,7 +319,7 @@ def sample_positions_colors_and_yaw_of_objects(
         for obj in objects:
             key = obj.name
             pos = get_position_of_object(obj)
-            color = get_color_of_drone(obj)
+            color = get_final_color_of_drone(obj)
             rotation = get_xyz_euler_rotation_of_object(obj)
             trajectories[key].append(Point4D(time, *pos))
             lights[key].append(